fix: Windows SQLite file locks and cross-platform test compatibility
- cache.go: add Store.Close() method; close DB on migration failure
- cli/cache.go: defer store.Close() in all cache commands
- cache_test.go/extra: defer store.Close() in tests
- config: move defaultDirFor tests to paths_unix_test.go (//go:build !windows)
  because filepath.IsAbs("/path") returns false on Windows
